+++ docs-site/sidebars.js
@@ -19,6 +19,7 @@ const sidebars = {
         'guides/browser-lifecycle',
         'guides/parallel',
         'guides/smart-locator',
+        'guides/semantic-locators',
         'guides/download-manager',
         'guides/console-errors',
         'guides/precondition',